write experiment metadata as attrs in HDF5 when TimelapseTiler.run is invoked
When the standard pipeline (scripts/pipeline_test.py) is used,
experiment metadata from *Acq.txt and *log.txt files are now included in
the HDF5 files produced for each imaging position.

Experiment metadata is included in the *Acq.txt and *log.txt files.
This metadata contains information relevant to the pipeline, such as the
channel info, spatial positions, and camera type.  The information
contains data that may be relevant for post-processing later, such as
pump info.

This feature takes advantage of the logfile_parser repository, which
parses log files to produce dictionaries.  HDF5 offers attributes for
each group for metadata, but the object type that can be included is
limited -- thus, I decided to convert the datetime objects e.g.
experiment start time, to UNIX timestamps (obvious solution) and
flattened dictionaries e.g. tables.  The alternative would be converting
dictionaries into datasets, but ultimately this is a design judgement
call.  In any case, I have been able to devise a way to deal with all
object types, so there should be some robustness in the code.  However,
there was an issue with Writer, i.e. I couldn't get it to write to the
root group of the HDF5 file, so I had to define a new 'metadata' group
-- which goes against my design intention.

I tested the new functionality with the unit test (experiment ID 19993).

This commit addresses issue #28 ('Store experiment metadata in HDF5
output').