write experiment metadata as attrs of root group rather than 'metadata'
Previously metadata was written to a 'metadata' group (newly created if
it hadn't previously existed) of the HDF5 file.  Now the metadata is
written to the root group of the HDF5 file.

This is a cosmetic change without a change to functionality.  Writing
metadata to the root group falls within our design paradigm of having
the attrs describe the group it is associated with.

* core/io/writer.py: Previously added a new numbered group if path was
'/'.  This was because the conditionals, particularly surrounding the
use of rsplit, didn't consider the case in which path was '/'.  Fixed it
by adding a conditional to specifically check for path == '/' and bypass
everything so that path remains '/' and write_meta() writes to the
correct path, i.e. root group.

I tested the new functionality with the unit test (experiment ID 19993).

This commit addresses the issue raised in a366c44f and issue #28 ('Store
experiment metadata in HDF5).  Issue closed.