Template-less trap finding.
Implemented a texture-based segmentation for finding traps.
It relies on finding the areas of the bright field images with the
highest entropy.
The hyperparameters will likely need to be optimised for various
set-ups.
This may fail if the device is overloaded or there is a lot of
contamination/dust in the images as those will be seen as traps.
Could be used as a first step, with an additional template-based trap
finding step afterwards.

Step one of #34