Simultaneous Lineaer and Folded Edgemasks
Processing run on the same data at the same time but stored in two
differnt ways to see the improvement of saving the edgemasks in
n_cells*n_tps array. Speed greatly improved (extraction is no longer
linear with time points or cell number), but at large cost in storage
space.
Related to #36