Use metadata and chunking to reduce file size.
Now that the metadata is stored in the HDF5 file, we read out the number
of timepoints from the acq file to fix the max time points for the
Baby Writer. This decreases the size of the HDF5 file significantly.
Additionally, I added a fixed chunk size for edge masks that takes the
full mask, a single time point, and an (arbitraily chosen) number of
cells (25). This also reduces the file size as it is a more efficient
chunk than the auto-chunked feature.
The folded file is now the same size as (if not smaller than) the
linear file. The improvements in speed of extraction have not changed.

This is another step forward in #36