Remove identified traps that are too close to the edge of the image
Segmentation (tiler) failed if the initial image processed has a trap
that is too close to the edge of the image -- i.e. less than half the
tile size away from the edge.

When a user segments experiment 19995, position 16 ('phl_ura8_016') --
either in isolation or in parallel with other positions -- the pipeline
writes metadata, but then fails at the initial time point (time point
0).  A `LinAlgError` is raised  (`numpy.linalg.LinAlgError:
SVD did not converge`).

`segment_traps()` in `pcore/traps.py` -- which identifies the location
of traps within an image -- assumes that all traps are a sufficient
distance away from the edge of the image.  This was not the case for
experiment 19995, position 16, time point 0, brightfield, z stack 0, in
which a trap was identified to be at (1039, 33) in a (1200, 1200) image
-- it is too near the bottom of the image as the y-coordinate (33) is
less than half of the tile size (96).  This trap happens to be the one
with the smallest minor axis (`ma`), deemed to be most likely a trap
without a cell.  As a result, when a box is drawn around the trap
location to be used as a trap template (i.e. when `template` is defined
in line 78), it goes out of the bounds of the image.

I fixed the issue by re-defining `traps`.  This variable now excludes
traps that are less than half a tile size away from the image edge.

I re-ran the pipeline on position 16 and the first few time points are
now segmented without error.

This bug was caused by 36171adff857632e17cc4fc18ef479e3de54839a, when
`traps` was last defined.  This commit fixes issue #45.