autoreg process outputs optimal AR order into a dataframe
'autoreg' process computed the closed-form solution of periodograms for
each time series in the input DataFrame.  This process required finding
the optimal AR order by minimising the AIC.  In this implementation, the
process now outputs the optimal AR order for each time series into a
third output DataFrame.  The order is the most important variable,
therefore it is useful to have that as part of the main output as well.

I extended the use of named tuples and leveraged the structural changes
in 713b49a095e707525f200f4e4705b9acf646b8ca.

* postprocessor/core/processes/autoreg.py: Added the additional output,
  and updated docstrings to reflect this.
* tests/test_autoreg.py: Added a variable to reflect the additional
  output.

I tested using a custom script on dummy sinusoids to confirm that code
changes don't affect the frequency and power dataframes (and generally,
the other functionality), and that the order dataframe makes sense.