mi process can seed train/test splitting in bootstraps
mi process produces a different output each time it is run.  This is
expected behaviour owing to the stochastic nature of computing mutual
information.  But it can be undesirable if the user wants to reproduce
the output -- particularly important for code testing.

I added an additional parameter to specify whether to seed
train_test_split() for reproducibility.  I want the seed to be different
for each iteration during bootstrapping -- otherwise, there would be no
point in bootstrapping as calculating mutual information relies on each
result being different.  So, I used the 'i' counter variable, which was
going to change in every iteration anyway, to seed train_test_split().

These changes thus allow me to test whether the output array is as
expected, rather than just whether the process crashes.  Therefore, the
testing is more 'proper' as it is more specific.