Fix logic of evaluating whether Crange & gammarange is None
When user invokes mi.run(), it raises a ValueError: The truth value of
an array with more than one element in ambiguous.

This is likely because later versions of numpy are more stringent about
using a.any() or a.all() that the versions MIdecoding or the mi
post-process was originally developed on.