Mean plotter no longer has built-in scaling
Users must now scale the input before passing it through the plotter.
In other words:
   PREVIOUSLY
      ax = mean_plot(trace_df, *args, scale=True)
   NOW
      from postprocessor.core.processes.standardscaler import standardscaler
      trace_scale = standardscaler.as_function(trace_df)
      ax = mean_plot(trace_scale, *args)

This creates slightly more burden on the user, but (hopefully) less
burden on the developer, for the following reasons:
- This avoids a 'multi-function function'.  In other words, the mean
  plotter used to scale the dataframe, THEN create a plot.  This should
  be two functions, and because core processes have `.as_function()`
  methods, scaling should be easy for the user.
- The mean plotter thus is more consistent with the other
  two-dimensional plotting routines, i.e. `single_plot` and
  `single_birth_plot`.  Refactoring to reduce code repetition should
  then be easier.

I didn't do the same thing for `heatmap` because scaling is intertwined
with setting the colormap scale.  That is, if `scale` is True, then the
axes must be even.  The handler *must* know whether the data has been
scaled in order to do so, so that is why I left scaling in `heatmap`.

This change is not backwards-compatible and users must check that their
data is scaled before invoking the mean plotter.