<fix>(writer.py): <Removed check of data type befor writing timepoint column in writer>
WHY IS THIS CHANGE NEEDED?:
With numpy version 1.21.6 the pipeline was failing to complete because the statement at line 567 of writer.py:
 if (
                df.columns.dtype == np.int
                or df.columns.dtype == np.dtype("uint")
                or df.columns.name == "timepoint"
            ):
was false - this was due to a mismatch between the numpy and pandas integer data types. This might be fixed if the correct version of numpy is used. On windows this seems to be affected by the installation of omero-py

HOW DOES THE CHANGE SOLVE THE PROBLEM?:
The if statement is removed so the timepoint column is always written

EVIDENCE THAT COMMIT WORKS:
Whole pipeline now works with unit test on Windows 10 (sce-bio-c04115)